+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # CHANGELOG
 
+## March 9, 2021 - 3.4.0
+
+QueryMembers - Added support for pagination by user_id [0c41232](https://github.com/GetStream/stream-chat-js/commit/0c412321bc4de81b123574041e0abadf89f235df)
+
+## March 9, 2021 - 3.3.0
+
+### Feature
+
+- Added the `Client.getRateLimits` function to retrieve rate limit quotas and usage, with the option to filter per platform and endpoint [#631](https://github.com/GetStream/stream-chat-js/pull/631)
+- Support reactions sync for pinned messages in channel state [#629](https://github.com/GetStream/stream-chat-js/pull/629)
+
+## March 8, 2021 - 3.2.0
+
+### Feature
+
+- Added the `Client.queryBannedUsers` function to query banned users with optional filters [#625](https://github.com/GetStream/stream-chat-js/pull/625)
+
 ## March 5, 2021 - 3.1.4
 
 ### Fix